jingyuan-api-python/requirements.txt
aiofiles==0.8.0
aiohttp==3.8.3
aiopathlib==0.5.0
aioredis==2.0.1
aioshutil==1.3
aiosignal==1.3.1
alembic==1.8.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.5.2
async-timeout==4.0.2
asyncmy==0.2.8
attrs==22.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.2.0
crcmod==1.7
cryptography==38.0.3
dnspython==2.2.1
ecdsa==0.18.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.103.2
frozenlist==1.3.3
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
jmespath==0.10.0
loguru==0.7.2
Mako==1.2.4
MarkupSafe==2.1.1
motor==3.1.1
multidict==6.0.2
openai==0.27.4
openpyxl==3.0.10
orjson==3.8.2
oss2==2.16.0
passlib==1.7.4
Pillow==9.3.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.6.0
pymongo==4.3.3
PyMySQL==1.0.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
redis==5.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
SQLAlchemy-Utils==0.38.3
SSIM-PIL==1.0.14
starlette==0.27.0
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.7.1
ua-parser==0.16.1
urllib3==1.26.12
user-agents==2.2.0
uvicorn==0.21.0
watchgod==0.8.2
websockets==10.4
win32-setctime==1.1.0
XlsxWriter==3.0.3
yarl==1.8.1
zipp==3.10.0
zope.event==4.5.0
zope.interface==5.5.2
tencentcloud-sdk-python==3.0.1074
qiniu==7.13.0